authentication and authorization using jwt

+++ controllers/api/v1/posts_api.js
@@ -19,37 +19,33 @@ module.exports.index= async function(req,res)
     })
 }
 
-module.exports.destroyPost=async function(req,res)
-{
-   
-    try
-    {
-        let post=await Post.findById(req.params.id);
-        // if(post.user==req.user.id)
-        // {
-            await post.remove();
-            await Comment.deleteMany({post:req.params.id});
-            
-            //req.flash("success","Post and comments deleted!");
-            //return res.redirect("back");
+module.exports.destroy = async function(req, res){
 
-            return res.json(200,{
-                message:"Posts and associated comments deleted successfully!"
+    try{
+        console.log("hello");
+        let post = await Post.findById(req.params.id);
+        console.log(post);
+        if (post.user == req.user.id){
+            post.remove();
+
+            await Comment.deleteMany({post: req.params.id});
+
+
+    
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully!"
             });
-        // }
-        // else
-        // {
-        //     req.flash("error","You are not associated to delete the post");
-        //     return res.redirect("back");
-        // }
-    }
-    catch(err)
-    {
-        console.log("*******",err);
-        //req.flash("error",err);
-        //return res.redirect("back");
-        return res.json(500,{
-            message:"Internal Server Error"
+        }else{
+            return res.json(401, {
+                message: "You cannot delete this post!"
+            });
+        }
+
+    }catch(err){
+        console.log('********', err);
+        return res.json(500, {
+            message: "Internal Server Error"
         });
     }
+    
 }
